Adding Markovify IPYNB File
This is my IPYNB documentation for the Markovify portion of this project! After running it a number of times and seeing what kinds of output comes with the length being extended, I decided on 500 lines in order to give GPT-2 the best chance to have a good body of work to work with. As you'll see, this code uses the Text-Conglomerate.txt file that I've included in this repository.